+++ ts/StatblockEditor.ts
@@ -15,21 +15,31 @@ module ImprovedInitiative {
     
     SaveCreature = () => {
       if(this.editorType() === 'advanced') {
-        var statBlockFromJSON = JSON.parse(this.statBlockJson());
+        try {
+            var statBlockFromJSON = JSON.parse(this.statBlockJson());
+        } catch (error) {
+            alert(`Couldn't parse JSON from advanced editor.`);
+            return;
+        }
         this.StatBlock($.extend(StatBlock.Empty(), statBlockFromJSON))
       }
       var editedStatBlock = this.StatBlock();
-      editedStatBlock.HP.Value = this.parseInt(editedStatBlock.HP.Value, 1)
-      editedStatBlock.AC.Value = this.parseInt(editedStatBlock.AC.Value, 10)
-      editedStatBlock.Abilities.Dex = this.parseInt(editedStatBlock.Abilities.Dex, 10)
+      if(this.editorType() === 'basic') {
+        editedStatBlock.HP.Value = this.parseInt(editedStatBlock.HP.Value, 1)
+        editedStatBlock.AC.Value = this.parseInt(editedStatBlock.AC.Value, 10)
+        editedStatBlock.Abilities.Dex = this.parseInt(editedStatBlock.Abilities.Dex, 10)
+      }
       this.saveCallback(editedStatBlock);
       this.StatBlock(null);
     }
     
     DeleteCreature = () => {
         var statBlock = this.StatBlock();
-        this.deleteCallback(statBlock.Player == 'player' ? Store.PlayerCharacters : Store.Creatures, statBlock.Id);
-        this.StatBlock(null);
+        if(confirm(`Delete statblock for ${statBlock.Name}? This cannot be undone.`))
+        {
+            this.deleteCallback(statBlock.Player == 'player' ? Store.PlayerCharacters : Store.Creatures, statBlock.Id);
+            this.StatBlock(null);
+        }
     }
     
     private parseInt: (value, defaultValue?: number) => number = (value, defaultValue: number = null) => {
